requirements.txt
Django==1.8.2
djangorestframework==3.1.2
geopy==1.10.0
psycopg2==2.6
